+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS2RRJ5N6PRJ7FS0WKB43J15"
+        "01KS37FSSBEQBXK28Z342V8YNH"
     )
     r"""The unique identifier of the data source"""