+++ packages/orq-rc/src/orq_ai_sdk/sdk.py
@@ -147,7 +147,9 @@ class Orq(BaseSDK):
         ), "The provided async_client must implement the AsyncHttpClient protocol."
 
         security: Any = None
-        if callable(api_key):
+        if api_key is None:
+            security = None
+        elif callable(api_key):
             # pylint: disable=unnecessary-lambda-assignment
             security = lambda: models_.Security(api_key=api_key())
         else: